Changed embedding logic and added new RBF kernel - Causal LLMs now round the attention mask to 0/1. This is now possible, as the gradient is not passed anyway - added some assertions. - Added debug statements in generator outputting the size of the layers. - Added a new and more efficient RBF kernel fixing the exploding memory during back pass. - Limited the amount of subspaces for text and token VMMD to 10 due to runtime problems. - Added more cli-update statements - Removed weight-decay leading to no collapse anymore.

+++ src/text/Embedding/LLM/causal_llm.py
@@ -86,7 +86,7 @@ class CausalLLM(HuggingModel, ABC):
                 torch.backends.cuda.enable_flash_sdp(False)
                 outputs = self.model(inputs_embeds=input_embeds,
                                      #attention_mask=causal_mask,
-                                     attention_mask=mask,
+                                     attention_mask=mask.int().float(),
                                      use_cache=False,
                                      output_hidden_states=True)
             embeddings = outputs.hidden_states
